// frontend/src/shared/lib/api/model/decompositionRunResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { DecompositionRunResponseErrorMessage } from './decompositionRunResponseErrorMessage';

export interface DecompositionRunResponse {
  run_id: string;
  status: string;
  rgroup_labels: string[];
  matched_count: number;
  unmatched_count: number;
  total_count: number;
  error_message?: DecompositionRunResponseErrorMessage;
}